NXT-126 Refactor resource list pages
