--- web/src/components/ProductResourcesListPage/ProductResourcesListPage.tsx
@@ -1,57 +1,0 @@
-import { NextSeo } from "next-seo";
-import { type FC } from "react";
-
-import { Breadcrumbs, Breadcrumb } from "@nice-digital/nds-breadcrumbs";
-
-import { Link } from "@/components/Link/Link";
-import { ProductHorizontalNav } from "@/components/ProductHorizontalNav/ProductHorizontalNav";
-import { ProductPageHeading } from "@/components/ProductPageHeading/ProductPageHeading";
-import { ResourceList } from "@/components/ResourceList/ResourceList";
-
-import { type ProductResourcesListPageProps } from "./ProductResourcesListPage.getServerSideProps";
-
-export const ProductResourcesListPage: FC<ProductResourcesListPageProps> = ({
-	resourceTypeName,
-	resourceGroups,
-	productPath,
-	product,
-	hasToolsAndResources,
-	hasInfoForPublicResources,
-	hasEvidenceResources,
-	hasHistory,
-}) => {
-	return (
-		<>
-			<NextSeo
-				title={`${resourceTypeName} | ${product.id} | Indicators | Standards and Indicators`}
-			/>
-
-			<Breadcrumbs>
-				<Breadcrumb to="/">Home</Breadcrumb>
-				<Breadcrumb to="/standards-and-indicators">
-					Standards and Indicators
-				</Breadcrumb>
-				<Breadcrumb to="/standards-and-indicators/indicators">
-					Indicators
-				</Breadcrumb>
-				<Breadcrumb to={productPath} elementType={Link}>
-					{product.id}
-				</Breadcrumb>
-				<Breadcrumb>{resourceTypeName}</Breadcrumb>
-			</Breadcrumbs>
-
-			<ProductPageHeading product={product} />
-
-			<ProductHorizontalNav
-				productTypeName="Indicator"
-				productPath={productPath}
-				hasToolsAndResources={hasToolsAndResources}
-				hasInfoForPublicResources={hasInfoForPublicResources}
-				hasEvidenceResources={hasEvidenceResources}
-				hasHistory={hasHistory}
-			/>
-
-			<ResourceList title={resourceTypeName} groups={resourceGroups} />
-		</>
-	);
-};
